front: rename globalMap reducer by commonMap
