+++ front/src/applications/editor/tools/pointEdition/components/PointEditionLayers.tsx
@@ -16,8 +16,8 @@ import type { EditorEntity } from 'applications/editor/typesEditorEntity';
 import { GeoJSONs, EditorSource, SourcesDefinitionsIndex } from 'common/Map/Layers';
 import { colors } from 'common/Map/theme';
 import { useInfraID } from 'common/osrdContext';
+import { useMapSettings } from 'reducers/commonMap';
 import { getEditorState } from 'reducers/editor/selectors';
-import { useMapSettings } from 'reducers/globalMap';
 import { NULL_GEOMETRY } from 'types';
 
 type BasePointEditionLayersProps = {
