+++ front/src/applications/editor/tools/rangeEdition/electrification/ElectrificationEditionLayers.tsx
@@ -22,8 +22,8 @@ import type { ExtendedEditorContextType } from 'applications/editor/types';
 import { GeoJSONs, SourcesDefinitionsIndex } from 'common/Map/Layers';
 import { colors } from 'common/Map/theme';
 import { useInfraID } from 'common/osrdContext';
+import { useMapSettings } from 'reducers/commonMap';
 import { getEditorState } from 'reducers/editor/selectors';
-import { useMapSettings } from 'reducers/globalMap';
 import { useAppDispatch } from 'store';
 
 export const ElectrificationEditionLayers = () => {
