+++ front/src/applications/editor/tools/selection/components/SelectionLayers.tsx
@@ -10,8 +10,8 @@ import type { ExtendedEditorContextType } from 'applications/editor/types';
 import { GeoJSONs } from 'common/Map/Layers';
 import { colors } from 'common/Map/theme';
 import { useInfraID } from 'common/osrdContext';
+import { useMapSettings } from 'reducers/commonMap';
 import { getEditorState } from 'reducers/editor/selectors';
-import { useMapSettings } from 'reducers/globalMap';
 import type { Zone } from 'types';
 import { zoneToFeature } from 'utils/mapHelper';
