+++ front/src/applications/editor/tools/trackSplit/components/TrackSplitLayers.tsx
@@ -16,8 +16,8 @@ import type { ExtendedEditorContextType } from 'applications/editor/types';
 import { GeoJSONs } from 'common/Map/Layers';
 import { colors } from 'common/Map/theme';
 import { useInfraID } from 'common/osrdContext';
+import { useMapSettings } from 'reducers/commonMap';
 import { getEditorState } from 'reducers/editor/selectors';
-import { useMapSettings } from 'reducers/globalMap';
 
 import type { TrackSplitState } from '../types';
 import { isOffsetValid } from '../utils';
