+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/helpers/__tests__/formatTimetableItemPayload.spec.ts
@@ -5,7 +5,7 @@ import type {
   TrainCategory,
   TrainMainCategory,
 } from 'common/api/osrdEditoastApi';
-import { defaultMapSettings } from 'reducers/globalMap';
+import { defaultMapSettings } from 'reducers/commonMap';
 import type {
   AddedExceptionId,
   IndexedOccurrenceId,
