+++ front/src/common/Map/Layers/OSMLayers/OSMLayers.tsx
@@ -2,7 +2,7 @@ import OpenStreetMapSource from 'common/Map/Sources/OpenStreetMap';
 import TerrainSource from 'common/Map/Sources/Terrain';
 import { colors } from 'common/Map/theme';
 import { LAYER_GROUPS_ORDER, LAYERS } from 'config/layerOrder';
-import type { MapStyle } from 'reducers/globalMap/types';
+import type { MapStyle } from 'reducers/commonMap/types';
 
 import Background from './Background';
 import Hillshade from '../Hillshade';
