+++ front/src/common/Map/Search/MapSearch.tsx
@@ -7,8 +7,8 @@ import MapSearchLine from 'common/Map/Search/MapSearchLine';
 import MapSearchOperationalPoint from 'common/Map/Search/MapSearchOperationalPoint';
 import MapSearchSignal from 'common/Map/Search/MapSearchSignal';
 import Tabs from 'common/Tabs';
-import { useMapSettingsActions } from 'reducers/globalMap';
-import type { MapSettings, Viewport } from 'reducers/globalMap/types';
+import { useMapSettingsActions } from 'reducers/commonMap';
+import type { MapSettings, Viewport } from 'reducers/commonMap/types';
 import { useAppDispatch } from 'store';
 
 import MapModalHeader from '../MapModalHeader';
