+++ front/src/common/Map/Settings/MapSettingsLayers.tsx
@@ -10,7 +10,7 @@ import SignalsSVGFile from 'assets/pictures/layersicons/layer_signal.svg';
 import OPsSVGFile from 'assets/pictures/layersicons/ops.svg';
 import SwitchesSVGFile from 'assets/pictures/layersicons/switches.svg';
 import SwitchSNCF, { SWITCH_TYPES } from 'common/BootstrapSNCF/SwitchSNCF';
-import { useMapSettings, useMapSettingsActions } from 'reducers/globalMap';
+import { useMapSettings, useMapSettingsActions } from 'reducers/commonMap';
 import type { MapState } from 'reducers/map';
 import { useAppDispatch } from 'store';
