+++ front/src/reducers/commonMap/index.ts
