+++ front/src/reducers/commonMap/types.ts
